+++ orc-c++/src/orc/RLEv1.cc
@@ -24,8 +24,8 @@
 
 namespace orc {
 
-const unsigned long MINIMUM_REPEAT = 3;
-const unsigned long BASE_128_MASK = 0x7f;
+const uint64_t MINIMUM_REPEAT = 3;
+const uint64_t BASE_128_MASK = 0x7f;
 
 signed char RleDecoderV1::readByte() {
   if (bufferStart == bufferEnd) {
@@ -40,24 +40,24 @@ signed char RleDecoderV1::readByte() {
   return *(bufferStart++);
 }
 
-unsigned long RleDecoderV1::readLong() {
-  unsigned long result = 0;
-  int offset = 0;
+uint64_t RleDecoderV1::readLong() {
+  uint64_t result = 0;
+  int64_t offset = 0;
   signed char ch = readByte();
   if (ch >= 0) {
-    result = static_cast<unsigned long>(ch);
+    result = static_cast<uint64_t>(ch);
   } else {
-    result = static_cast<unsigned long>(ch) & BASE_128_MASK;
+    result = static_cast<uint64_t>(ch) & BASE_128_MASK;
     while ((ch = readByte()) < 0) {
       offset += 7;
-      result |= (static_cast<unsigned long>(ch) & BASE_128_MASK) << offset;
+      result |= (static_cast<uint64_t>(ch) & BASE_128_MASK) << offset;
     }
-    result |= static_cast<unsigned long>(ch) << (offset + 7);
+    result |= static_cast<uint64_t>(ch) << (offset + 7);
   }
   return result;
 }
 
-void RleDecoderV1::skipLongs(unsigned long numValues) {
+void RleDecoderV1::skipLongs(uint64_t numValues) {
   while (numValues > 0) {
     if (readByte() >= 0) {
       --numValues;
@@ -68,15 +68,15 @@ void RleDecoderV1::skipLongs(unsigned long numValues) {
 void RleDecoderV1::readHeader() {
   signed char ch = readByte();
   if (ch < 0) {
-    remainingValues = static_cast<unsigned long>(-ch);
+    remainingValues = static_cast<uint64_t>(-ch);
     repeating = false;
   } else {
-    remainingValues = static_cast<unsigned long>(ch) + MINIMUM_REPEAT;
+    remainingValues = static_cast<uint64_t>(ch) + MINIMUM_REPEAT;
     repeating = true;
     delta = readByte();
     value = isSigned
         ? unZigZag(readLong())
-        : static_cast<long>(readLong());
+        : static_cast<int64_t>(readLong());
   }
 }
 
@@ -100,16 +100,16 @@ void RleDecoderV1::seek(PositionProvider& location) {
   skip(location.next());
 }
 
-void RleDecoderV1::skip(unsigned long numValues) {
+void RleDecoderV1::skip(uint64_t numValues) {
   while (numValues > 0) {
     if (remainingValues == 0) {
       readHeader();
     }
-    unsigned long count = std::min(numValues, remainingValues);
+    uint64_t count = std::min(numValues, remainingValues);
     remainingValues -= count;
     numValues -= count;
     if (repeating) {
-      value += delta * static_cast<long>(count);
+      value += delta * static_cast<int64_t>(count);
     } else {
       skipLongs(count);
     }
@@ -117,9 +117,9 @@ void RleDecoderV1::skip(unsigned long numValues) {
 }
 
 void RleDecoderV1::next(int64_t* const data,
-                        const unsigned long numValues,
+                        const uint64_t numValues,
                         const char* const notNull) {
-  unsigned long position = 0;
+  uint64_t position = 0;
   // skipNulls()
   if (notNull) {
     // Skip over null values.
@@ -133,41 +133,41 @@ void RleDecoderV1::next(int64_t* const data,
       readHeader();
     }
     // How many do we read out of this block?
-    unsigned long count = std::min(numValues - position, remainingValues);
-    unsigned long consumed = 0;
+    uint64_t count = std::min(numValues - position, remainingValues);
+    uint64_t consumed = 0;
     if (repeating) {
       if (notNull) {
-        for (unsigned long i = 0; i < count; ++i) {
+        for (uint64_t i = 0; i < count; ++i) {
           if (notNull[position + i]) {
-            data[position + i] = value + static_cast<long>(consumed) * delta;
+            data[position + i] = value + static_cast<int64_t>(consumed) * delta;
             consumed += 1;
           }
         }
       } else {
-        for (unsigned long i = 0; i < count; ++i) {
-          data[position + i] = value + static_cast<long>(i) * delta;
+        for (uint64_t i = 0; i < count; ++i) {
+          data[position + i] = value + static_cast<int64_t>(i) * delta;
         }
         consumed = count;
       }
-      value += static_cast<long>(consumed) * delta;
+      value += static_cast<int64_t>(consumed) * delta;
     } else {
       if (notNull) {
-        for (unsigned long i = 0 ; i < count; ++i) {
+        for (uint64_t i = 0 ; i < count; ++i) {
           if (notNull[i]) {
             data[position + i] = isSigned
                 ? unZigZag(readLong())
-                : static_cast<long>(readLong());
+                : static_cast<int64_t>(readLong());
             ++consumed;
           }
         }
       } else {
         if (isSigned) {
-          for (unsigned long i = 0; i < count; ++i) {
+          for (uint64_t i = 0; i < count; ++i) {
             data[position + i] = unZigZag(readLong());
           }
         } else {
-          for (unsigned long i = 0; i < count; ++i) {
-            data[position + i] = static_cast<long>(readLong());
+          for (uint64_t i = 0; i < count; ++i) {
+            data[position + i] = static_cast<int64_t>(readLong());
           }
         }
         consumed = count;
