+++ orc-c++/src/orc/RLEv2.hh
@@ -42,19 +42,19 @@ public:
   /**
   * Seek over a given number of values.
   */
-  void skip(unsigned long numValues) override;
+  void skip(uint64_t numValues) override;
 
   /**
   * Read a number of values into the batch.
   */
-  void next(int64_t* data, unsigned long numValues,
+  void next(int64_t* data, uint64_t numValues,
             const char* notNull) override;
 
 private:
 
   // Used by PATCHED_BASE
   void adjustGapAndPatch() {
-    curGap = static_cast<unsigned long>(unpackedPatch[patchIdx]) >>
+    curGap = static_cast<uint64_t>(unpackedPatch[patchIdx]) >>
       patchBitSize;
     curPatch = unpackedPatch[patchIdx] & patchMask;
     actualGap = 0;
@@ -64,7 +64,7 @@ private:
     while (curGap == 255 && curPatch == 0) {
       actualGap += 255;
       ++patchIdx;
-      curGap = static_cast<unsigned long>(unpackedPatch[patchIdx]) >>
+      curGap = static_cast<uint64_t>(unpackedPatch[patchIdx]) >>
         patchBitSize;
       curPatch = unpackedPatch[patchIdx] & patchMask;
     }
@@ -97,15 +97,15 @@ private:
   return result;
 }
 
-  long readLongBE(uint32_t bsz);
+  int64_t readLongBE(uint64_t bsz);
   int64_t readVslong();
   uint64_t readVulong();
   uint64_t readLongs(int64_t *data, uint64_t offset, uint64_t len,
                      uint64_t fb, const char* notNull = nullptr) {
-  unsigned long ret = 0;
+  uint64_t ret = 0;
 
   // TODO: unroll to improve performance
-  for(unsigned long i = offset; i < (offset + len); i++) {
+  for(uint64_t i = offset; i < (offset + len); i++) {
     // skip null positions
     if (notNull && !notNull[i]) {
       continue;
@@ -123,10 +123,10 @@ private:
     // handle the left over bits
     if (bitsLeftToRead > 0) {
       result <<= bitsLeftToRead;
-      bitsLeft -= bitsLeftToRead;
+      bitsLeft -= static_cast<uint32_t>(bitsLeftToRead);
       result |= (curByte >> bitsLeft) & ((1 << bitsLeftToRead) - 1);
     }
-    data[i] = static_cast<long>(result);
+    data[i] = static_cast<int64_t>(result);
     ++ret;
   }
 
@@ -148,24 +148,24 @@ private:
 
   unsigned char firstByte;
   uint64_t runLength;
-  unsigned long runRead;
+  uint64_t runRead;
   const char *bufferStart;
   const char *bufferEnd;
-  long deltaBase; // Used by DELTA
-  unsigned int byteSize; // Used by SHORT_REPEAT and PATCHED_BASE
-  long firstValue; // Used by SHORT_REPEAT and DELTA
-  long prevValue; // Used by DELTA
+  int64_t deltaBase; // Used by DELTA
+  uint64_t byteSize; // Used by SHORT_REPEAT and PATCHED_BASE
+  int64_t firstValue; // Used by SHORT_REPEAT and DELTA
+  int64_t prevValue; // Used by DELTA
   uint32_t bitSize; // Used by DIRECT, PATCHED_BASE and DELTA
   uint32_t bitsLeft; // Used by anything that uses readLongs
   uint32_t curByte; // Used by anything that uses readLongs
   uint32_t patchBitSize; // Used by PATCHED_BASE
-  unsigned long unpackedIdx; // Used by PATCHED_BASE
-  unsigned long patchIdx; // Used by PATCHED_BASE
-  long base; // Used by PATCHED_BASE
-  unsigned long curGap; // Used by PATCHED_BASE
-  long curPatch; // Used by PATCHED_BASE
-  long patchMask; // Used by PATCHED_BASE
-  long actualGap; // Used by PATCHED_BASE
+  uint64_t unpackedIdx; // Used by PATCHED_BASE
+  uint64_t patchIdx; // Used by PATCHED_BASE
+  int64_t base; // Used by PATCHED_BASE
+  uint64_t curGap; // Used by PATCHED_BASE
+  int64_t curPatch; // Used by PATCHED_BASE
+  int64_t patchMask; // Used by PATCHED_BASE
+  int64_t actualGap; // Used by PATCHED_BASE
   DataBuffer<int64_t> unpacked; // Used by PATCHED_BASE
   DataBuffer<int64_t> unpackedPatch; // Used by PATCHED_BASE
 };
